refactor cachable interface. Because rx2 has multiple types of observables give common interface all implementations can extend.
