+++ shelf/src/main/java/com/toddway/shelf/rx/Cacheable.kt
@@ -1,0 +1,7 @@
+package com.toddway.shelf.rx
+
+internal interface Cacheable<T> {
+
+    var cache: T?
+    val isCacheValid: Boolean
+}
