+++ shelf/src/main/java/com/toddway/shelf/rx/Shelfable.kt
@@ -1,16 +1,12 @@
 package com.toddway.shelf.rx
 
-
 import com.toddway.shelf.ShelfItem
-import io.reactivex.Maybe
-import io.reactivex.Single
-import io.reactivex.SingleTransformer
 
 /**
  * Created by tway on 4/24/16.
  * updated nschwermann 12/21/17
  */
-class Shelfable<T>(observeNew: Single<T>, internal var item: ShelfItem, internal var type: Class<T>) : RxCacheable<T>(observeNew.toMaybe()) {
+internal class Shelfable<T>(internal var item: ShelfItem, internal var type: Class<T>) : Cacheable<T> {
 
     override var cache: T?
         get() = item.get(type)
@@ -25,14 +21,4 @@ class Shelfable<T>(observeNew: Single<T>, internal var item: ShelfItem, internal
         return item
     }
 
-    companion object {
-
-        internal fun<T> observeCache(item: ShelfItem, type: Class<T>) : Maybe<T> {
-            return Shelfable(Maybe.empty<T>().toSingle(), item, type).observeCacheIfValid()
-        }
-
-        internal fun<T> cacheOrNew(item: ShelfItem, type: Class<T>): SingleTransformer<T, T> {
-            return SingleTransformer { observable -> Shelfable(observable, item, type).observeCacheOrNew().toSingle() }
-        }
-    }
 }
